Use floating pointer values in combined elimination script
Before this was using integer values, however for some results such
as time or energy it make sense to have floating point values here.

MAGEEC itself still assumes integer values in the results, and this
needs to be updated to use floating point values instead.

ChangeLog:

	* script/combined-elimination.py: Treat result values as floating
	point values instead of integers.

+++ script/combined-elimination.py
@@ -239,7 +239,7 @@ def build_and_measure(src_dir, build_dir, install_dir, build_system,
 
         src_file = values[0]
         module_name = values[2]
-        result = int(values[6])
+        result = float(values[6])
 
         if src_file in result_files:
             print ('-- Duplicate results for module ' + module_name)
